# internal/admin/api/openapi.yaml
openapi: "3.0.3"
info:
  title: GoProxify Admin API
  description: API REST du Control Plane GoProxify
  version: "1.0"
servers:
  - url: /
    description: Instance courante
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gpx_session
    bearerAuth:
      type: http
      scheme: bearer
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [nodes]
      summary: Approuver / révoquer un agent
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [approved, revoked]
      responses:
        "200":
          description: OK
    delete:
      tags: [nodes]
      summary: Supprimer un agent
      responses:
        "204":
          description: Supprimé
...